+++ Tapjoy/TapjoyInterstitialCustomEvent.m
@@ -2,8 +2,10 @@
 #import "TapjoyInterstitialCustomEvent.h"
 #import <Tapjoy/TJPlacement.h>
 #import <Tapjoy/Tapjoy.h>
-#import "MPLogging.h"
-#import "MoPub.h"
+#if __has_include("MoPub.h")
+    #import "MPLogging.h"
+    #import "MoPub.h"
+#endif
 
 @interface TapjoyInterstitialCustomEvent () <TJPlacementDelegate>
 @property (nonatomic, strong) TJPlacement *placement;
@@ -35,7 +37,7 @@
         NSMutableDictionary *connectOptions = [[NSMutableDictionary alloc] init];
         [connectOptions setObject:@(enableDebug) forKey:TJC_OPTION_ENABLE_LOGGING];
         [self setupListeners];
-
+        
         [Tapjoy connect:sdkKey options:connectOptions];
 
         self.isConnecting = YES;
@@ -54,7 +56,7 @@
     if (self.isConnecting) {
         return;
     }
-
+    
     // Attempt to establish a connection to Tapjoy
     if (![Tapjoy isConnected]) {
         [self initializeWithCustomNetworkInfo:info];
@@ -122,6 +124,7 @@
 - (void)tjcConnectSuccess:(NSNotification*)notifyObj {
     MPLogInfo(@"Tapjoy connect Succeeded");
     self.isConnecting = NO;
+    [self fetchMoPubGDPRSettings];
     [self requestPlacementContent];
 
     [[NSNotificationCenter defaultCenter] removeObserver:self];
@@ -131,6 +134,24 @@
     MPLogInfo(@"Tapjoy connect Failed");
     self.isConnecting = NO;
     [[NSNotificationCenter defaultCenter] removeObserver:self];
+}
+
+// Collect latest MoPub GDPR settings and pass them to Tapjoy
+-(void)fetchMoPubGDPRSettings {
+    // If the GDPR applies setting is unknown, assume it has been skipped/unset
+    MPBool gdprApplies = [MoPub sharedInstance].isGDPRApplicable;
+    if (gdprApplies != MPBoolUnknown ) {
+        //Turn the MPBool into a proper bool
+        if(gdprApplies == MPBoolYes) {
+            [Tapjoy subjectToGDPR:YES];
+            
+            NSString *consentString = [[MoPub sharedInstance] canCollectPersonalInfo] ? @"1" : @"0";
+            [Tapjoy setUserConsent: consentString];
+        } else {
+            [Tapjoy subjectToGDPR:NO];
+            [Tapjoy setUserConsent:@"-1"];
+        }
+    }
 }
 
 @end