+++ Vungle/MPVungleRouter.m
@@ -6,12 +6,13 @@
 //
 
 #import "MPVungleRouter.h"
-#import "MPInstanceProvider+Vungle.h"
-#import "MPLogging.h"
+#if __has_include("MoPub.h")
+    #import "MPLogging.h"
+    #import "MPRewardedVideoError.h"
+    #import "MPRewardedVideo.h"
+    #import "MoPub.h"
+#endif
 #import "VungleInstanceMediationSettings.h"
-#import "MPRewardedVideoError.h"
-#import "MPRewardedVideo.h"
-#import "MoPub.h"
 
 static NSString *const VunglePluginVersion = @"6.2.0";
 
@@ -54,7 +55,12 @@ typedef NS_ENUM(NSUInteger, SDKInitializeState) {
 }
 
 + (MPVungleRouter *)sharedRouter {
-    return [[MPInstanceProvider sharedProvider] sharedMPVungleRouter];
+    static MPVungleRouter * sharedRouter;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        sharedRouter = [[MPVungleRouter alloc] init];
+    });
+    return sharedRouter;
 }
 
 - (void)initializeSdkWithInfo:(NSDictionary *)info {