+++ AdMob/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 7.30.0.2
+    * update adapters to remove dependency on MPInstanceProvider
+    * Update adapters to be compatible with MoPub iOS SDK framework
+
   * 7.30.0.1
   	* Updated the adapter's cocoapods dependency to MoPub version 5.0