+++ Flurry/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 8.4.0.3  
+    * update adapters to remove dependency on MPInstanceProvider
+    * Update adapters to be compatible with MoPub iOS SDK framework
+
   * 8.4.0.2
   	* Updated the adapter's cocoapods dependency to MoPub version 5.0